add: Modify merge to insertion sort
A practical measure to speed up development;
temporarily change implementation to
simplest and inefficient algo to be able to
finish the rest of the project.

Ideally the sorting algo can be easily adapted
to something more efficient once its interface
is clearer. This ought to help with my
organisation, too.