feat(wip): Add chunk partitioning
This is the first half of a new divide and conquer
algo. A naive split by 10 (with ops for testing
more efficient sizes later) is made on init and
every node is assigned a "chunk".

This commit currently splits the stack
correctly according to pretty good manual testing.

It became clear that the compromise with mergesort
which i thought was a chunk sort, is actually
a partitioning by chunks with greedy merging from
stack b to a. It should be sorted in one pass back
and forth overall.

TODO: Unit tests, acceptance tests, the merge,
the rotations can be interleaved for probably
significant cost savings. Consider finding a way
to use swaps. Consider light presorting during
splitting.