ref: Refactor error handling
ps_error() now cleans up and exits immediately
because upon error, no further processing should
be done.

In some places it was more convenient to keep
the previous soft failure trace in static
functions, but this would hide the exact source
of an error. They were left in part because these
functions are extensively tried and tested; like
those of init.c.

A possible better solution would be to let
clean_up hold a static reference to the stacks
so that it can blitz everythig at any point.
Quick ideas for it would have its inititalisation
be kind of ugly, so i've left it for now.